# contracts/openapi/ticket-service.v1.yaml
openapi: 3.1.0
info:
  title: Ticket Service API
  version: "1.0.0"
  description: >-
    Regulatory registry and appeal card API of the MFO Appeals Platform. Covers manual
    registration, card updates, classification, comments, and search (TASK_01B). Conventions
    (docs/05): base path /api/v1, JSON with camelCase fields, ISO-8601 UTC timestamps, RFC 7807
    Problem Details for errors, X-Correlation-ID propagation, cursorless page/pageSize pagination,
    optimistic locking via an expectedVersion field, and Idempotency-Key for create commands.
    Status, stage, and assignment are Flowable projections and are not editable through this API.
servers:
  - url: /api/v1
    description: Versioned base path (mounted by the service).
security:
  - bearerAuth: []
...
paths:
  /tickets:
    post:
      tags: [tickets]
      operationId: createManualTicket
      summary: Register an appeal manually.
      description: >-
        Registers any written appeal, allocates a unique registration number, and stores the
        applicant (and optional representative). Emits ticket.created.v1 via the transactional
        outbox. Supply Idempotency-Key to make retries safe: idempotency is scoped to the
        authenticated caller, so a repeated key from the same caller with the same payload returns
        the original ticket, a different payload yields 409, and a legacy (pre-scoping) key yields a
        409 reconciliation-required. Registering isConfidential=true requires a confidentiality-
        cleared role (SUPERVISOR/OMBUDSMAN); otherwise 403.
      parameters:
        - $ref: '#/components/parameters/CorrelationId'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTicketRequest'
      responses:
        '201':
          description: The appeal was registered.
          headers:
            X-Correlation-ID:
              $ref: '#/components/headers/CorrelationId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TicketResponse'
        '200':
          description: An appeal with the same Idempotency-Key already exists; the original is returned.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TicketResponse'
        '400':
          $ref: '#/components/responses/Problem'
        '401':
          $ref: '#/components/responses/Problem'
        '403':
          $ref: '#/components/responses/Problem'
        '409':
          $ref: '#/components/responses/Problem'
        '422':
          $ref: '#/components/responses/Problem'
    get:
      tags: [tickets]
      operationId: searchTickets
      summary: Search appeals.
      description: >-
        Filters appeals by registration number, national identifier (IIN/BIN), full name, contract
        number, status, stage, product, classifier, channel, assignee, team, and received/registered
        date ranges. Results are paginated.
      parameters:
        - $ref: '#/components/parameters/CorrelationId'
        - {name: registrationNumber, in: query, schema: {type: string}, description: Exact registration number.}
        - {name: identifierValue, in: query, schema: {type: string}, description: Exact national identifier (IIN/BIN) of an attached party.}
        - {name: fullName, in: query, schema: {type: string}, description: Case-insensitive partial match on a party's full name.}
        - {name: contractNumber, in: query, schema: {type: string}, description: Exact related contract number.}
        - {name: statusCode, in: query, schema: {type: string}, description: Current status code.}
        - {name: stageCode, in: query, schema: {type: string}, description: Current stage code.}
        - {name: productCode, in: query, schema: {type: string}, description: Credit-product code.}
        - {name: classifierCode, in: query, schema: {type: string}, description: Question-classifier code.}
        - {name: channelCode, in: query, schema: {type: string}, description: Intake channel code.}
        - {name: assigneeId, in: query, schema: {type: string, format: uuid}, description: Current assignee identifier.}
        - {name: teamId, in: query, schema: {type: string, format: uuid}, description: Current team identifier.}
        - {name: receivedFrom, in: query, schema: {type: string, format: date-time}, description: Inclusive lower bound on receivedAt.}
        - {name: receivedTo, in: query, schema: {type: string, format: date-time}, description: Inclusive upper bound on receivedAt.}
        - {name: registeredFrom, in: query, schema: {type: string, format: date-time}, description: Inclusive lower bound on registeredAt.}
        - {name: registeredTo, in: query, schema: {type: string, format: date-time}, description: Inclusive upper bound on registeredAt.}
        - {name: page, in: query, schema: {type: integer, minimum: 1, default: 1}, description: 1-based page number.}
        - {name: pageSize, in: query, schema: {type: integer, minimum: 1, maximum: 100, default: 20}, description: Page size.}
      responses:
        '200':
          description: A page of matching appeals.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedTickets'
        '400':
          $ref: '#/components/responses/Problem'
  /tickets/{ticketId}:
    parameters:
      - $ref: '#/components/parameters/TicketId'
    get:
      tags: [tickets]
      operationId: getTicket
      summary: Get an appeal by identifier.
      parameters:
        - $ref: '#/components/parameters/CorrelationId'
      responses:
        '200':
          description: The appeal card.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TicketResponse'
        '404':
          $ref: '#/components/responses/Problem'
    patch:
      tags: [tickets]
      operationId: updateTicketDetails
      summary: Update editable appeal-card details.
      description: >-
        Updates card text fields (subject, description, source channel, contract number). Status,
        stage, and assignment are not editable here (they change only through the Flowable
        projection). Requires expectedVersion for optimistic locking; a mismatch yields 409. Emits
        ticket.updated.v1.
      parameters:
        - $ref: '#/components/parameters/CorrelationId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateTicketRequest'
      responses:
        '200':
          description: The updated appeal card.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TicketResponse'
        '404':
          $ref: '#/components/responses/Problem'
        '409':
          $ref: '#/components/responses/Problem'
        '422':
          $ref: '#/components/responses/Problem'
  /tickets/{ticketId}/access:
    parameters:
      - $ref: '#/components/parameters/TicketId'
    get:
      tags: [tickets]
      operationId: getTicketAccess
      summary: Report the caller's capabilities on an appeal.
      description: >-
        Returns this service's own data-scope decision for the authenticated caller: whether they may
        read the appeal and whether they may mutate it. Exposed so another service can enforce this
        policy without duplicating it and without reading this database (ADR-0008, ADR-004) — the
        Document Service checks canMutate before attaching or moving evidence, because inferring that
        right from a successful read would let a controlled read/audit role lend its broad scope to a
        mutation. The probe has no side effects and writes no audit record. An appeal outside the
        caller's scope and an appeal that does not exist both answer false to everything, so the
        response never reveals whether an appeal exists.
      parameters:
        - $ref: '#/components/parameters/CorrelationId'
      responses:
        '200':
          description: The caller's read and mutation capabilities on this appeal.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TicketAccessResponse'
        '401':
          $ref: '#/components/responses/Problem'
        '403':
          $ref: '#/components/responses/Problem'
  /tickets/{ticketId}/classify:
    parameters:
      - $ref: '#/components/parameters/TicketId'
    post:
      tags: [tickets]
      operationId: classifyTicket
      summary: Set the appeal classification.
      description: >-
        Sets product, question classifier, and priority. Requires expectedVersion for optimistic
        locking. Emits ticket.classified.v1.
      parameters:
        - $ref: '#/components/parameters/CorrelationId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ClassifyRequest'
      responses:
        '200':
          description: The reclassified appeal card.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TicketResponse'
        '404':
          $ref: '#/components/responses/Problem'
        '409':
          $ref: '#/components/responses/Problem'
        '422':
          $ref: '#/components/responses/Problem'
  /tickets/{ticketId}/decision:
    parameters:
      - $ref: '#/components/parameters/TicketId'
    post:
      tags: [tickets]
      operationId: recordDecision
      summary: Record the decision on an appeal.
      description: >-
        Records the decision code, text, author, and timestamp. Requires expectedVersion for
        optimistic locking. Emits ticket.decision_recorded.v1.
      parameters:
        - $ref: '#/components/parameters/CorrelationId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RecordDecisionRequest'
      responses:
        '200':
          description: The appeal card with the recorded decision.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TicketResponse'
        '404':
          $ref: '#/components/responses/Problem'
        '409':
          $ref: '#/components/responses/Problem'
        '422':
          $ref: '#/components/responses/Problem'
  /tickets/{ticketId}/close:
    parameters:
      - $ref: '#/components/parameters/TicketId'
    post:
      tags: [tickets]
      operationId: closeTicket
      summary: Close an appeal.
      description: >-
        Closes an appeal after validating that a decision, decision text, responsible employee,
        decision date, a response date or a justified absence of response, and a closure reason are
        present (docs/01). Sets the retention date (at least five years) and the terminal status.
        A missing prerequisite yields 422. Emits ticket.closed.v1.
      parameters:
        - $ref: '#/components/parameters/CorrelationId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CloseTicketRequest'
      responses:
        '200':
          description: The closed appeal card.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TicketResponse'
        '404':
          $ref: '#/components/responses/Problem'
        '409':
          $ref: '#/components/responses/Problem'
        '422':
          $ref: '#/components/responses/Problem'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >-
        Signed access token issued by the IAM service. The ticket service verifies it independently
        (signature, fixed algorithm allowlist, issuer, audience, expiry) and enforces its own
        permission and data-scope policy on every request, even when called directly (not via the
        BFF). Missing/invalid tokens yield 401 with a WWW-Authenticate: Bearer challenge; missing
        permission or data scope yields 403.
  parameters:
    TicketId:
      name: ticketId
      in: path
      required: true
      schema: {type: string, format: uuid}
      description: Internal ticket identifier (UUIDv7).
    CorrelationId:
      name: X-Correlation-ID
      in: header
      required: false
      schema: {type: string}
      description: Correlation identifier propagated across services; generated when absent.
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: false
      schema: {type: string, maxLength: 200}
      description: Client-supplied key making a create retry-safe.
  headers:
    CorrelationId:
      schema: {type: string}
      description: Correlation identifier for this request.
  responses:
    Problem:
      description: An RFC 7807 Problem Details error.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
...